requirements.txt
google-generativeai>=0.3.1
python-dotenv>=1.0.1